Labøvelse - Oppgave 1
Error: The first child option element of a select element with a required attribute, and without a multiple attribute, and without a size attribute whose value is greater than 1, must have either an empty value attribute, or must have no text content. Consider either adding a placeholder option label, or adding a size attribute with a value equal to the number of option elements.

From line 59, column 25; to line 59, column 46

          <option value="norge">Norge<